Prevent failure to create SLURL when an object chats but is not on the object list.
Fixes an issue I couldn't remember nor find on the issue tracker.

SLURLs for found objects remain unchanged.
SLURLs for objects that are not found, but we could find their owners now display the owner's position and region, with "?owner_not_object" appended to the slurl as a note to the user
SLURLs for objects that are not found, and we could not find their owners, now display as being from the position 0, 0, 0 in the users region