Fix ClickToWalk not allowing mouse-walk (behavior change)
Ansariel and worked together on this to solve http://jira.phoenixviewer.com/browse/FIRE-15189

The behavior change is that when your own avatar is clicked and the mouse is dragged, the camera no longer simply turns on a single axis,
since the behavior for dragging from anywhere else is still the old single-axis pan, this is not a very big deal. I can even see this being more useful!