[Radar] Correct a flaw in which initialization of the avatar list would just be a refresh of the gui
Without first populating the list, this was pretty much nothing doing, now the list will populate initially.
Previously, it could be that opening and closing the radar (when not in announce even when closed mode) would yield an empty radar should the sim have been stingy or the connection been rather latent.