Remove all selected entries from asset blacklist
Previously, the "Remove" button in the asset blacklist floater removed
the last clicked-on entry only - even when selecting multiple rows (by
holding CTRL and clicking on entries).
This behaviour is unintuitive and tedious when trying to selectively
remove multiple entries.

The feature was requested in the group chat. Tested both removing a
single entry and multiple entries with two different avatars on Linux.