Do not steal focus from LSL script error console
When selecting a list item in the script error console the focus was
moved to the script text editor. Errors could only be copied with the
keyboard shortcut while clicking and holding the mouse on a single item.

The output can now be selected to allow copying with keyboard shortcuts
(CTRL-C) and (multiple) selection like any other list (CTRL + left mouse
or CTRL-A).