Disable vertical tab resizing over maximum width
In some situations, the label of tabs changes which caused the tab to
resize.
For example, when using vertical tabs for instant messages in combination
with "Display Names (w/Username)" for chat tabs, the display name does not
instantly load sometimes (when not in the avatar name cache?) and the
username is set. After the display name loads the tab label is set to the
display name which triggered a resizing of the tab and could cause it to
overlap the chat box.