Couple things~
Fixes for crash on startup.
Fixes Friti's issue of llceflib not downloading on linux
Changes USESYSTEMLIBS to STANDALONE in cmake files so STANDALONE possibly might work as expected
Fixes a few compiler warnings.
Imports a fix for win7 and sync llapp.cpp a bit.
Fixes the cmake issue with VIEWER_CHANNEL_NOSPACES (Thanks for helping, Friti!)